Move manage project to the left of project name
