+++ libs/ui-compose-ext/src/main/kotlin/mono/ui/compose/components/Icons.kt
@@ -68,16 +68,18 @@ object Icons {
         Svg(
             attrs = {
                 size(iconSize, iconSize)
-                viewBox(512, 512)
-                fill("currentColor")
+                viewBox(800, 800)
+                fill("none")
             }
         ) {
             SvgPath(
-                "M121 32c-29 0-55 20-62 49L2 308l-2 16v92c0 35 29 64 64 64h384c35 0 64-29 64-64v-92l-2-16-57-227c-7-29-33-49-62-49H121Zm0 64h270l48 192h-51c-12 0-23 7-29 18l-14 28c-5 11-17 18-29 18H196c-12 0-23-7-29-18l-14-28c-5-11-17-18-29-18H73l48-192Z"
-            )
-            SvgPath(
-                "M142 146c0-10 8-18 18-18h196a17 17 0 1 1 0 35H160c-10 0-18-8-18-17ZM123 229c0-10 8-18 18-18h232a17 17 0 1 1 0 35H141c-10 0-18-8-18-17Z"
-            )
+                "M771 442 644 145c-7-17-24-28-43-28H199c-19 0-36 11-43 28L29 442m742 0v186c0 51-42 93-93 93H122c-51 0-93-42-93-93V442m742 0H586c-26 0-47 21-47 47v46c0 26-21 47-46 47H307c-25 0-46-21-46-47v-46c0-26-21-47-46-47H29m242-208 244-2m-291 94h337"
+            ) {
+                attr("stroke", "currentColor")
+                attr("stroke-linecap", "round")
+                attr("stroke-linejoin", "round")
+                attr("stroke-width", "50")
+            }
         }
     }
 }
