+++ src/main/sass/style/header/working-file.scss
@@ -15,7 +15,6 @@
     .file-info {
         display: flex;
         align-items: center;
-        margin-left: 2px;
         height: 100%;
         padding: 6px 4px 6px 8px;
         font-family: $monospaceFont;
