+++ src/main/sass/style/variables.scss
@@ -13,7 +13,8 @@ $keyboard-shortcuts-zindex: 200;
 $export-text-zindex: 201;
 $edit-text-zindex: 202;
 $dropdown-menu-zindex: 203;
-$tooltip-zindex: 204;
 $rename-project-modal-zindex: 205;
 $all-projects-modal-zindex: 500;
 $remove-project-modal-zindex: 501;
+
+$tooltip-zindex: 10000;
